Moved class GreetingService to a separate file
