+++ BetterConsoleApp/ConsoleUI/Program.cs
@@ -4,7 +4,6 @@ using System.IO;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
-using Microsoft.Extensions.Logging;
 
 using Serilog;
 
@@ -40,27 +39,6 @@ namespace ConsoleUI
                    .AddJsonFile("appsettings.json", false, true)
                    .AddJsonFile($"appsettings.{Environment.GetEnvironmentVariable("ASPNETCORE_ENVIRONMENT") ?? "Production"}.json", true)
                    .AddEnvironmentVariables();
-        }
-    }
-
-    public class GreetingService
-    {
-        private readonly ILogger<GreetingService> log;
-
-        private readonly IConfiguration config;
-
-        public GreetingService(ILogger<GreetingService> log, IConfiguration config)
-        {
-            this.log=log;
-            this.config=config;
-        }
-
-        public void Run()
-        {
-            for (int i = 0; i < config.GetValue<int>("LoopTimes"); i++)
-            {
-                log.LogInformation("Run number { runNumber }", i);
-            }
         }
     }
 }
